Removed warning
git-svn-id: https://svn.wxwidgets.org/svn/wx/wxWidgets/trunk@23171 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775

+++ samples/widgets/notebook.cpp
@@ -136,7 +136,7 @@ protected:
 
     // is the value in range?
     bool IsValidValue(int val) const
-        { return (val >= 0) && (val < m_notebook->GetPageCount()); }
+        { return (val >= 0) && (val < (int) m_notebook->GetPageCount()); }
 
     // the controls
     // ------------
